Add tests for tls with host name override
